+++ Sources/BucketQueue/TestHistory/Model/TestEntryHistoryId.swift
@@ -5,33 +5,14 @@ import QueueModels
 import RunnerModels
 
 public struct TestEntryHistoryId: Hashable {
-    public let testEntry: TestEntry
-    public let testDestination: TestDestination
-    public let toolResources: ToolResources
-    public let buildArtifacts: BuildArtifacts
     public let bucketId: BucketId
+    public let testEntry: TestEntry
     
     public init(
-        testEntry: TestEntry,
-        testDestination: TestDestination,
-        toolResources: ToolResources,
-        buildArtifacts: BuildArtifacts,
-        bucketId: BucketId
+        bucketId: BucketId,
+        testEntry: TestEntry
     ) {
-        self.testEntry = testEntry
-        self.testDestination = testDestination
-        self.toolResources = toolResources
-        self.buildArtifacts = buildArtifacts
         self.bucketId = bucketId
-    }
-    
-    public init(testEntry: TestEntry, bucket: Bucket) {
-        self.init(
-            testEntry: testEntry,
-            testDestination: bucket.testDestination,
-            toolResources: bucket.toolResources,
-            buildArtifacts: bucket.buildArtifacts,
-            bucketId: bucket.bucketId
-        )
+        self.testEntry = testEntry
     }
 }