+++ Sources/EmceeLib/Commands/DumpRuntimeTestsCommand.swift
@@ -87,7 +87,7 @@ public final class DumpRuntimeTestsCommand: Command {
                     environment: testArgFileEntry.environment,
                     numberOfRetries: testArgFileEntry.numberOfRetries
                 ),
-                testRunnerTool: testArgFileEntry.toolResources.testRunnerTool,
+                testRunnerTool: testArgFileEntry.testRunnerTool,
                 testTimeoutConfiguration: testTimeoutConfigurationForRuntimeDump,
                 testsToValidate: testArgFileEntry.testsToRun,
                 xcTestBundleLocation: testArgFileEntry.buildArtifacts.xcTestBundle.location