+++ Sources/EmceeLib/Utils/DefaultSimulatorControllerProvider.swift
@@ -11,19 +11,25 @@ import fbxctest
 
 public final class DefaultSimulatorControllerProvider: SimulatorControllerProvider {
     private let additionalBootAttempts: UInt
+    private let automaticSimulatorShutdown: TimeInterval
     private let developerDirLocator: DeveloperDirLocator
     private let simulatorBootQueue: DispatchQueue
+    private let simulatorOperationTimeouts: SimulatorOperationTimeouts
     private let simulatorStateMachineActionExecutorProvider: SimulatorStateMachineActionExecutorProvider
     
     public init(
         additionalBootAttempts: UInt,
+        automaticSimulatorShutdown: TimeInterval,
         developerDirLocator: DeveloperDirLocator,
         simulatorBootQueue: DispatchQueue,
+        simulatorOperationTimeouts: SimulatorOperationTimeouts,
         simulatorStateMachineActionExecutorProvider: SimulatorStateMachineActionExecutorProvider
     ) {
         self.additionalBootAttempts = additionalBootAttempts
+        self.automaticSimulatorShutdown = automaticSimulatorShutdown
         self.developerDirLocator = developerDirLocator
         self.simulatorBootQueue = simulatorBootQueue
+        self.simulatorOperationTimeouts = simulatorOperationTimeouts
         self.simulatorStateMachineActionExecutorProvider = simulatorStateMachineActionExecutorProvider
     }
 
@@ -33,23 +39,21 @@ public final class DefaultSimulatorControllerProvider: SimulatorControllerProvid
         testDestination: TestDestination,
         testRunnerTool: TestRunnerTool
     ) throws -> SimulatorController {
-        return StateMachineDrivenSimulatorController(
-            additionalBootAttempts: additionalBootAttempts,
-            bootQueue: simulatorBootQueue,
-            developerDir: developerDir,
-            developerDirLocator: developerDirLocator,
-            simulatorOperationTimeouts: SimulatorOperationTimeouts(
-                create: 30,
-                boot: 180,
-                delete: 20,
-                shutdown: 20
-            ),
-            simulatorStateMachine: SimulatorStateMachine(),
-            simulatorStateMachineActionExecutor: try simulatorStateMachineActionExecutorProvider.simulatorStateMachineActionExecutor(
-                simulatorControlTool: simulatorControlTool,
-                testRunnerTool: testRunnerTool
-            ),
-            testDestination: testDestination
+        return ActivityAwareSimulatorController(
+            automaticShutdownTimePeriod: automaticSimulatorShutdown,
+            delegate: StateMachineDrivenSimulatorController(
+                additionalBootAttempts: additionalBootAttempts,
+                bootQueue: simulatorBootQueue,
+                developerDir: developerDir,
+                developerDirLocator: developerDirLocator,
+                simulatorOperationTimeouts: simulatorOperationTimeouts,
+                simulatorStateMachine: SimulatorStateMachine(),
+                simulatorStateMachineActionExecutor: try simulatorStateMachineActionExecutorProvider.simulatorStateMachineActionExecutor(
+                    simulatorControlTool: simulatorControlTool,
+                    testRunnerTool: testRunnerTool
+                ),
+                testDestination: testDestination
+            )
         )
     }
 }