+++ Sources/QueueModels/TestEntryConfiguration.swift
@@ -9,39 +9,42 @@ public struct TestEntryConfiguration: Codable, CustomStringConvertible, Hashable
     public let buildArtifacts: BuildArtifacts
     public let developerDir: DeveloperDir
     public let pluginLocations: Set<PluginLocation>
+    public let simulatorControlTool: SimulatorControlTool
     public let simulatorSettings: SimulatorSettings
     public let testDestination: TestDestination
     public let testEntry: TestEntry
     public let testExecutionBehavior: TestExecutionBehavior
+    public let testRunnerTool: TestRunnerTool
     public let testTimeoutConfiguration: TestTimeoutConfiguration
     public let testType: TestType
-    public let toolResources: ToolResources
 
     public init(
         buildArtifacts: BuildArtifacts,
         developerDir: DeveloperDir,
         pluginLocations: Set<PluginLocation>,
+        simulatorControlTool: SimulatorControlTool,
         simulatorSettings: SimulatorSettings,
         testDestination: TestDestination,
         testEntry: TestEntry,
         testExecutionBehavior: TestExecutionBehavior,
+        testRunnerTool: TestRunnerTool,
         testTimeoutConfiguration: TestTimeoutConfiguration,
-        testType: TestType,
-        toolResources: ToolResources
+        testType: TestType
     ) {
         self.buildArtifacts = buildArtifacts
         self.developerDir = developerDir
         self.pluginLocations = pluginLocations
+        self.simulatorControlTool = simulatorControlTool
         self.simulatorSettings = simulatorSettings
         self.testDestination = testDestination
         self.testEntry = testEntry
         self.testExecutionBehavior = testExecutionBehavior
+        self.testRunnerTool = testRunnerTool
         self.testTimeoutConfiguration = testTimeoutConfiguration
         self.testType = testType
-        self.toolResources = toolResources
     }
     
     public var description: String {
-        return "<\(type(of: self)): \(testEntry) \(testType) \(testDestination) \(buildArtifacts) \(pluginLocations) \(simulatorSettings) \(testExecutionBehavior) \(testTimeoutConfiguration) \(toolResources) \(developerDir)>"
+        return "<\(type(of: self)): \(testEntry) \(testType) \(testDestination) \(buildArtifacts) \(pluginLocations) \(simulatorSettings) \(testExecutionBehavior) \(testTimeoutConfiguration) \(simulatorControlTool) \(testRunnerTool) \(developerDir)>"
     }
 }